RAG_Research_Assistant/Agentic_RAG/requirements.txt
langchain
langchain-community
streamlit
python-dotenv
openai
semanticscholar
chromadb
sentence-transformers
langchain_openai
transformers